at this point, I can do all windday shit, the crane stuff works, but it is unlinked from the project/user data structure, the resulting crane paths are shit, and it is still using the filesystem for all sort of business...
