+++ app/upload.py
@@ -22,16 +22,6 @@ def allowed_file(filename,extensions):
     return '.' in filename and \
            filename.rsplit('.', 1)[1] in extensions
 
-def clear_shpfiles():
-    shpdir = os.path.join(app.config['UPLOAD_FOLDER'], 'shapefiles')
-    if os.path.exists(shpdir):
-        for the_file in os.listdir(shpdir):
-            file_path = os.path.join(shpdir, the_file)
-            try:
-                if os.path.isfile(file_path):
-                    os.unlink(file_path)
-            except Exception, e:
-                print e
 
 @app.route('/api/windyday/<project_name>/upload', methods=['POST'])
 @auth.login_required
@@ -65,29 +55,4 @@ def upload_file(project_name):
             print 'seasonality plot = ' + str(t2-t1)
             print 'tmatrix = ' +str(t3-t2)
             print 'save = ' + str(t4-t3)
-            return flask.jsonify(result={"status": 200})
-            
-@app.route('/api/cranepath/zipupload',methods=['POST'])
-@auth.login_required
-def upload_ziplfile():
-    clear_shpfiles()
-    print "At upload"
-    if request.method == 'POST':
-        fileobj = request.files['file']
-        print "Got File"
-        if fileobj and allowed_file(fileobj.filename,ZIP):
-            print "File is allowed"
-            #filename = secure_filename(file.filename)
-            fileobj.save(os.path.join(app.config['UPLOAD_FOLDER'], 'shapefiles.zip'))
-            path = os.path.join(app.config['UPLOAD_FOLDER'], 'shapefiles')
-            with zipfile.ZipFile(os.path.join(app.config['UPLOAD_FOLDER'], 'shapefiles.zip'), "r") as z:
-                for member in z.namelist():
-                    filename = os.path.basename(member)
-                    if not filename: continue
-                    # copy file (taken from zipfile's extract)
-                    source = z.open(member)
-                    target = file(os.path.join(path, filename), "wb")
-                    with source, target:
-                        shutil.copyfileobj(source, target)
-                    
             return flask.jsonify(result={"status": 200})
