+++ gui/app/scripts/directives/piechart.js
@@ -15,12 +15,12 @@ angular.module('windopsApp')
       items: '='
     },
     controller: function ($scope, $element, $attrs) {
-      console.log(2);
+      //console.log(2);
 
     },
     template: '<div id="container" style="margin: 0 auto">not working</div>',
     link: function (scope, element, attrs) {
-      console.log(3);
+      //console.log(3);
       var chart = new Highcharts.Chart({
         chart: {
           renderTo: 'container',
